Fix issues with post tagging and isolate in it's own hook
