+++ projects/plugins/jetpack/extensions/blocks/blogging-prompt/edit.js
@@ -1,39 +1,19 @@
 import apiFetch from '@wordpress/api-fetch';
 import { InspectorControls, useBlockProps } from '@wordpress/block-editor';
 import { PanelBody, Spinner, ToggleControl, withNotices } from '@wordpress/components';
-import { useDispatch, useSelect } from '@wordpress/data';
 import { useEffect, useState } from '@wordpress/element';
-import { escapeHTML } from '@wordpress/escape-html';
 import { __, _x } from '@wordpress/i18n';
 import './editor.scss';
 import icon from './icon';
-
-// Tries to create a tag or fetch it if it already exists.
-// @link https://github.com/WordPress/gutenberg/blob/98b58d9042eda7590659c6cce2cf7916ba99aaa1/packages/editor/src/components/post-taxonomies/flat-term-selector.js#L55
-function findOrCreateTag( tagName ) {
-	const escapedTagName = escapeHTML( tagName );
-
-	return apiFetch( {
-		path: `/wp/v2/tags`,
-		method: 'POST',
-		data: { name: escapedTagName },
-	} ).catch( error => {
-		if ( error.code !== 'term_exists' ) {
-			return Promise.reject( error );
-		}
-
-		return Promise.resolve( {
-			id: error.data.term_id,
-			name: tagName,
-		} );
-	} );
-}
+import { usePromptTags } from './use-prompt-tags';
 
 function BloggingPromptsBetaEdit( { attributes, noticeOperations, noticeUI, setAttributes } ) {
 	const [ isLoading, setLoading ] = useState( true );
 	const { gravatars, prompt, promptId, showLabel, showResponses, tagsAdded } = attributes;
 	const blockProps = useBlockProps( { className: 'jetpack-blogging-prompts' } );
-	const { editPost } = useDispatch( 'core/editor' );
+
+	// Add the prompt tags to the post, if they haven't already been added.
+	usePromptTags( promptId, tagsAdded, setAttributes );
 
 	// Fetch the prompt by id, if present, otherwise the get the prompt for today.
 	useEffect( () => {
@@ -73,49 +53,6 @@ function BloggingPromptsBetaEdit( { attributes, noticeOperations, noticeUI, setA
 				noticeOperations.createErrorNotice( error.message );
 			} );
 	}, [ isLoading, noticeOperations, promptId, setAttributes, setLoading ] );
-
-	// Get the tags for the post.
-	const { tags, tagIds, tagsHaveResolved } = useSelect( select => {
-		const { getEditedPostAttribute } = select( 'core/editor' );
-		const { getEntityRecords, hasFinishedResolution } = select( 'core' );
-		const _termIds = getEditedPostAttribute( 'tags' );
-
-		const query = {
-			_fields: 'id,name',
-			context: 'view',
-			include: _termIds.join( ',' ),
-			per_page: -1,
-		};
-
-		return {
-			tagIds: _termIds,
-			tags: _termIds.length ? getEntityRecords( 'taxonomy', 'post_tag', query ) : [],
-			tagsHaveResolved: hasFinishedResolution( 'getEntityRecords', [
-				'taxonomy',
-				'post_tag',
-				query,
-			] ),
-		};
-	}, [] );
-
-	// Add the related prompt tags, if they haven't been added already.
-	useEffect( () => {
-		if ( tagsAdded || ! tagsHaveResolved || ! promptId || ! Array.isArray( tags ) ) {
-			return;
-		}
-
-		if ( ! tags.some( tag => tag.name && 'dailyprompt' === tag.name ) ) {
-			findOrCreateTag( 'dailyprompt' ).then( tag => {
-				editPost( { tags: [ ...tagIds, tag.id ] } );
-			} );
-		} else if ( ! tags.some( tag => tag.name && `dailyprompt-${ promptId }` === tag.name ) ) {
-			findOrCreateTag( `dailyprompt-${ promptId }` ).then( tag => {
-				editPost( { tags: [ ...tagIds, tag.id ] } );
-			} );
-		} else {
-			setAttributes( { tagsAdded: true } );
-		}
-	}, [ editPost, promptId, setAttributes, tagsAdded, tagsHaveResolved, tags, tagIds ] );
 
 	const onShowLabelChange = newValue => {
 		setAttributes( { showLabel: newValue } );
